Move closure implementation from ssm-mem.c to ssm-closure.c
